backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for ColombianStay"
authors = [
    {name = "JulianSpinoza",email = "juliancamilosimpson@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "pymongo (>=4.15.5,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
